dispatch: Raise SystemExit after catching wrapped errors
Previously, a wrapped error (eg. a CommandError) would result in `dispatch()`
returning successfully. This means we are powerless to affect program exit
after an error. The first step in being able to correctly exit non-zero in the face
of a wrapped error is to raise SystemExit instead of returning after a wrapped error.

This is quite a departure from the existing API, though in 99% of use cases it shouldn't matter.
The other thing to consider is that a parser failure or `--help` call already resulted in
a SystemExit from within argparse, so users expecting to do things after dispatch finishes
already needed to handle error cases.