Exit with status 1 (or CommandError.code) on wrapped exception
Instead of exiting with status 0 when a wrapped exception occurs, exit with status 1
to indicate failure.

The code can be customized by passing the `code` kwarg to `CommandError`.
You can use `code=0` to get the old behaviour.

This is a breaking change but the previous behaviour would have almost always been a bug.