docs: add Read the Docs config (#160)
The builds for PRs are successful but the latest/stable ones are failing. This needs to be stabilised.

As the RTD docs mention[1], "With builds from pull requests, only HTML formats are generated. Other formats are resource intensive and will be built after merging."

The least we can do to make sure that the up-to-date documentation is properly published is turn off PDF which is failing (possibly due to the badges in README.rst). Should anyone request building PDF, it can be re-enabled after testing and fixing.