fix: assumed prog name in tests (#171)
The tests are run both locally and in CI via tox.  This guarantees
reproducible runs.  However, they are ran differently in some other
conditions.  For instance, `python -m pytest` would result in a
different prog name and break tests where it was hard-coded.

This both fixes the tests and adds a tox env that would check this case
in the future.

Fixes #170