Exit failure on CommandError or other wrapped error (#161)
* dispatching: Fix some inaccurate docs around error handling

* dispatch: Raise SystemExit after catching wrapped errors

Previously, a wrapped error (eg. a CommandError) would result in `dispatch()`
returning successfully. This means we are powerless to affect program exit
after an error. The first step in being able to correctly exit non-zero in the face
of a wrapped error is to raise SystemExit instead of returning after a wrapped error.

This is quite a departure from the existing API, though in 99% of use cases it shouldn't matter.
The other thing to consider is that a parser failure or `--help` call already resulted in
a SystemExit from within argparse, so users expecting to do things after dispatch finishes
already needed to handle error cases.

* Refactor test code to allow easily reading out/err after SystemExit

Add exit status to CmdResult. Preserve existing run(exit=True) behaviour
to avoid needing to change every usage.

* Exit with status 1 (or CommandError.code) on wrapped exception

Instead of exiting with status 0 when a wrapped exception occurs, exit with status 1
to indicate failure.

The code can be customized by passing the `code` kwarg to `CommandError`.
You can use `code=0` to get the old behaviour.

This is a breaking change but the previous behaviour would have almost always been a bug.

* Fix tests for new CommandError behavior

Mark in tests that it is expected to exit with status 1.

Also add a case to test_command_error to test setting code.

* tutorial: Update section on CommandError

to clarify that it will always exit, and mention the option to control the exit status.

* style: fix lint errors

* docs: remove outdated warnings

---------

Co-authored-by: Andy Mikhailenko <neithere@gmail.com>